requirements.txt
winsdk
colorlog
aiohttp
python-dotenv